backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
supabase==2.24.0
python-dotenv==1.0.1
google-generativeai==0.8.3

